set up watch connectivity in a--delegate
